+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+
+* [#777](https://github.com/babylonlabs-io/finality-provider/pull/777) fix: align pubrand and proof indexing with actual block heights
+
 ## v2.0.2
 
 ### Bug fixes